[DOC] Update strong_type documentation

+++ test/snippet/core/detail/strong_type_error_window.cpp
@@ -1,11 +1,11 @@
 #include <seqan3/core/detail/strong_type.hpp>
 
-struct error : seqan3::detail::strong_type<unsigned, error>
+struct error : seqan3::detail::strong_type<uint8_t, error>
 {
-    using seqan3::detail::strong_type<unsigned, error>::strong_type;
+    using seqan3::detail::strong_type<uint8_t, error>::strong_type;
 };
 
-struct window_size : seqan3::detail::strong_type<unsigned, window_size>
+struct window_size : seqan3::detail::strong_type<uint8_t, window_size>
 {
-    using seqan3::detail::strong_type<unsigned, window_size>::strong_type;
+    using seqan3::detail::strong_type<uint8_t, window_size>::strong_type;
 };
